MaxSessions and MaxConnections are limited to 64k not 64.
In environments with large numbers of targets and/or paths
and/or initiators the number of sessions needed can be far greater
than 64.  Every tuplet of target/path/initiator needs a session.

The limit for istgt is 65536 for these values, (64k) not 64.

(cherry picked from commit 3962cc667d98be43d339fa8626ac02b3c654aa1b)